refactor rename package names to lowercase
