+++ nhais-adaptor/adaptor/outgoing/common/common_adaptor.py
